requirements.txt
Django==4.2
psycopg2-binary==2.9.6
djangorestframework==3.14.0
Pillow>=8.0.0
djangorestframework
